save-test

+++ src/com/apmath/loans/domain/models/loans/LoanForCalculationsInterface.kt
@@ -5,7 +5,7 @@ import com.apmath.loans.domain.data.Money
 interface LoanForCalculationsInterface {
     val amount: Money?
     val term: Int?
-    val interest: Float
+    val interest: Int
     val date: String
     val remainingTerm: Int?
     val rounding: Int
